pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "forecasting-system"
version = "0.1.0"
description = "Phase 1 scaffold for an event-driven financial forecasting system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.8",
    "openpyxl>=3.1",
    "pandas>=2.2",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
